+++ app/Http/Controllers/IncidentTypeController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\StoreIncidentTypeRequest;
 use App\Http\Requests\UpdateIncidentTypeRequest;
 use App\Models\IncidentType;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
 use Inertia\Inertia;
 
 class IncidentTypeController extends Controller
@@ -16,7 +17,7 @@ class IncidentTypeController extends Controller
 
         Inertia::flash('toast', ['type' => 'success', 'message' => __('Incident type created.')]);
 
-        return to_route('form-management.index');
+        return to_route('form-management.index', ['region_id' => $request->validated('region_id')]);
     }
 
     public function update(UpdateIncidentTypeRequest $request, IncidentType $incidentType): RedirectResponse
@@ -25,15 +26,17 @@ class IncidentTypeController extends Controller
 
         Inertia::flash('toast', ['type' => 'success', 'message' => __('Incident type updated.')]);
 
-        return to_route('form-management.index');
+        return to_route('form-management.index', ['region_id' => $incidentType->region_id]);
     }
 
-    public function destroy(IncidentType $incidentType): RedirectResponse
+    public function destroy(Request $request, IncidentType $incidentType): RedirectResponse
     {
+        abort_unless($request->user()->canAccessRegion($incidentType->region_id), 403);
+
         $incidentType->delete();
 
         Inertia::flash('toast', ['type' => 'success', 'message' => __('Incident type deleted.')]);
 
-        return to_route('form-management.index');
+        return to_route('form-management.index', ['region_id' => $incidentType->region_id]);
     }
 }
